run rustfmt

+++ src/simd.rs
@@ -37,7 +37,7 @@ use multiversion::multiversion;
 ))]
 pub fn simd_match_select<M, const N: usize>(pattern: [u8; N], mask: M, data: [u8; N]) -> bool
 where
-M: Into<u64>,
+    M: Into<u64>,
     LaneCount<N>: SupportedLaneCount,
 {
     Mask::from_bitmask(mask.into())
@@ -87,7 +87,6 @@ where
     M: Into<u64>,
     LaneCount<N>: SupportedLaneCount,
 {
-    (!Mask::from_bitmask(mask.into() as u64)
-        | Simd::from_array(data).simd_eq(Simd::from_array(pattern)))
-    .all()
+    (!Mask::from_bitmask(mask.into()) | Simd::from_array(data).simd_eq(Simd::from_array(pattern)))
+        .all()
 }
